Extract common fixtures in conftest.py
